Clean up API and add test data.
This removes Core Data from the application.  It would be too
complicated and confusing to get it to work with that -- it's simpler to
just implement it ourselves.  This will let people focus on learning
just Swift instead of Swift+Core Data.

In addition, the various references have been updated.  There is support
for Songs, Artists, and Albums.  You can get access the various elements
via the relational properties.  Note that artists must be accessed via
albums for songs for now, but that could change in the future.

In addition, the Library that is getting initialized has some default
data.  As long as the library object from the app delegate is used, the
application should seamlessly transition when the switch is made to use
the Music API-provided data.  (An importer needs to be written for
that.)