Update API for previous/current/upcoming songs.
This wasn't very well built in the previous commit.  This fixes it up
into something that is actually usable.

The general idea is as follows: there are history, current, and upcoming
variables, all of which contain QueueItem objects.  A QueueItem object
adds the necessary extra information to songs in the Queue (such as the
number of votes).

When advancing to the next song, the current song moves to the history,
and a new current song is taken from the front of upcoming.

There is also a notification that the rest of the application can use to
know when the queue changes.  This is how the UI should be updated when
the current song changes.