Improve AppleLibrary parsing.
(artist, album) is meant to be unique, and we weren't treating it as
such.  We were potentially associating albums with the same name that
have been released by different artists as the same thing.

In addition, there was one more spot where we were adding albums to an
artist twice.  That is now gone.