Fix UpNextTableViewController.
There was an issue with the number of rows reported for the table as
compared to the length of the array.

In addition, I added support for updating the view when the currently
playing song changes.  (It will reload all of the table view data.)

As a general word of advice, cellForRowAtIndexPath is meant to be fast,
and so it's generally good to not put much data acquisition code in that
method.