Make Queue work in terms of QueueItem objects.
This is how I had intended for it to be designed.  A Song no longer has
any idea as to how many votes it has received.  This is now stored
entirely within the QueueItem class.  This is needed for situations
where we look back at the QueueItem objects in the history when the
songs have been requests again and again.

A Song, however, does have a notion of whether or not it has been voted
for by the current user IN THE LIST OF UPCOMING SONGS.  So we clear the
"did vote for" indicator after a Song gets played.  We still need to
make this visually happen if a Song finishes while in the Library.

Currently, only upvotes are working, but I'll work to add support for
downvotes and tracking that sort of thing as well.

The code-level changes consisted mostly of these functionality changes
and also of cleanup to split many of the files into "Local" and "Remote"
variants.