Add support for sending "Codes".
This will allow us to communicate commands and state between the server
and client.  It will be used for everything from version validation to
noting when the library needs to be reset/is finished "loading".

This also rewrites some of the receive code in terms of the wonderful
extensions made to NSData and NSMutableData.  Overall, the networking
code becomes much easier to understand, and it also becomes much more
consistent.  Bugs shall be eliminated!