Update project metadata.
Restrict devices to the iPhone family.  (This is done mostly for the
demonstration.  We can re-enable the iPad later.)

Set the deployment target to iOS 9.0.  This ensures the best
compatibility with devices.

Restrict the orientation to portrait.  While some views look fine in
landscape mode, some of them look horrible.  Until someone audits all of
the views in the application, it's easier to just enable portrait.